putting-together/src/main/java/org/agoncal/fascicle/cdi/puttingtogether/BookService.java
package org.agoncal.fascicle.cdi.puttingtogether;

import javax.inject.Inject;

/**
 * @author Antonio Goncalves
 * http://www.antoniogoncalves.org
 * --
 */
// tag::adocsnippet[]
@Loggable
public class BookService {

  @Inject
  @ThirteenDigits
  private NumberGenerator numberGenerator;

  public Book createBook(String title, Float price, String description) {
    Book book = new Book(title, price, description);
    book.setNumber(numberGenerator.generateNumber());
    return book;
  }
}
// end::adocsnippet[]
